swaps the value of two argument
Create a function called swaps() that interchanges the values of the two arguments sent to it (pass these arguments by reference). Make the function into a template, so it can be used with all numerical data types( char, int, float and so on). Write a main() program to exercise the function with several types.