Constructor,Destructor,and newanddeleteOperator
Define a class matrix having the with data member row, col as integer and an pointer 
to int pointer. introduce constructor that initializes the row, col and also the pointer 
to null value. another parameterized constructor that should allocate the memory for the
matrix object as specified row and column by its arguments. Your class should include another
member  function for input the matrix elements and for display the matrix elements and a destructor
to clean up the memory. Use new and delete operators for allocate and deallocate the memory.